タイトルの画像貼り付けとスクリプトのフォルダーをまとめました。
時間関係は”Timer”に、ボタン関係は”Button”というフォルダーに入れました。
