+++ Yubi-Puzzle/Assets/Scenes/Result.unity
@@ -148,7 +148,7 @@ RectTransform:
   m_GameObject: {fileID: 93790223}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalScale: {x: 3, y: 3, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 1955077015}
@@ -406,10 +406,10 @@ RectTransform:
   m_Children: []
   m_Father: {fileID: 2003882951}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0.5, y: 0.5}
-  m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -296, y: 177}
-  m_SizeDelta: {x: 200, y: 50}
+  m_AnchorMin: {x: 0, y: 1}
+  m_AnchorMax: {x: 0, y: 1}
+  m_AnchoredPosition: {x: 200, y: -100}
+  m_SizeDelta: {x: 300, y: 100}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &746916229
 MonoBehaviour:
@@ -458,8 +458,8 @@ MonoBehaviour:
   m_faceColor:
     serializedVersion: 2
     rgba: 4294967295
-  m_fontSize: 36
-  m_fontSizeBase: 36
+  m_fontSize: 50
+  m_fontSizeBase: 50
   m_fontWeight: 400
   m_enableAutoSizing: 0
   m_fontSizeMin: 18
@@ -611,10 +611,10 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1243932146}
+  - component: {fileID: 1243932151}
   - component: {fileID: 1243932150}
   - component: {fileID: 1243932149}
   - component: {fileID: 1243932148}
-  - component: {fileID: 1243932147}
   m_Layer: 5
   m_Name: ToTitleButton
   m_TagString: Untagged
@@ -631,7 +631,7 @@ RectTransform:
   m_GameObject: {fileID: 1243932145}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalScale: {x: 3, y: 3, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 1498120731}
@@ -639,21 +639,9 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -0.38119507, y: -56}
+  m_AnchoredPosition: {x: 0, y: -200}
   m_SizeDelta: {x: 160, y: 30}
   m_Pivot: {x: 0.5, y: 0.5}
---- !u!114 &1243932147
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1243932145}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fdf5883421cc7414f8c3ef0df6526089, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &1243932148
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -736,6 +724,18 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1243932145}
   m_CullTransparentMesh: 1
+--- !u!114 &1243932151
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1243932145}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: fdf5883421cc7414f8c3ef0df6526089, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1498120730
 GameObject:
   m_ObjectHideFlags: 0
