+++ Yubi-Puzzle/Assets/Scripts/Button/StartButtomScript.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 75e891ccdcd93dd4c9f710e348b47419
