--- Yubi-Puzzle/Assets/Scripts/CloseButtonScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 806462e66f286854f91fd3f893ca365a
