--- Yubi-Puzzle/Assets/Scripts/CountdownScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9683b2e4e8c0b554e954b8b431f54670
