--- Yubi-Puzzle/Assets/Scripts/MoveTimerScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 30daabd44e54c6a479c969728136f5b1
