--- Yubi-Puzzle/Assets/Scripts/PuzzleTimerScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fd48a92647ab11f40b43185b6b714823
