--- Yubi-Puzzle/Assets/Scripts/RetryButtonScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f9056e6a9a38db44891838661c8478e0
