--- Yubi-Puzzle/Assets/Scripts/StartButtomScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 75e891ccdcd93dd4c9f710e348b47419
