+++ Yubi-Puzzle/Assets/Scripts/Timer.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b81567b00f64d884c91cafefa0b116ac
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
