--- Yubi-Puzzle/Assets/Scripts/ToTitleButtonScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fdf5883421cc7414f8c3ef0df6526089
